requirements.txt
streamlit>=1.37.0
pandas>=2.2.2
openpyxl>=3.1.5
XlsxWriter>=3.2.0
reportlab>=4.2.2
requests>=2.32.3
python-dotenv>=1.0.1
google-genai
python-dotenv
requests
